(spas:feature:layout) files renamed
