+++ src/spas/libs/layout/feature/src/lib/feature-layout.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'feature-layout',
-  templateUrl: './layout-feature.component.html',
-  styleUrls: ['./layout-feature.component.scss']
+  templateUrl: './feature-layout.component.html',
+  styleUrls: ['./feature-layout.component.scss']
 })
 export class LayoutFeatureComponent implements OnInit {
