+++ src/spas/libs/layout/feature/src/lib/feature-layout.module.spec.ts
@@ -1,5 +1,5 @@
 import { async, TestBed } from '@angular/core/testing';
-import { LayoutFeatureModule } from './layout-feature.module';
+import { LayoutFeatureModule } from './feature-layout.module';
 
 describe('LayoutFeatureModule', () => {
   beforeEach(async(() => {
