+++ src/spas/nx.json
@@ -44,7 +44,7 @@
     "layout-domain": {
       "tags": ["scope:layout", "type:domain"]
     },
-    "layout-feature": {
+    "feature-layout": {
       "tags": ["scope:layout", "type:feature"]
     },
     "shared-ui-material": {
